refactor: 로또를 판매하는 로직과 결과를 확인하는 로직을 분리해라
