+++ src/main/java/lotto/Application.java
@@ -2,12 +2,14 @@ package lotto;
 
 import lotto.config.AppConfig;
 import lotto.controller.LottoController;
-import lotto.service.LottoService;
+import lotto.service.LottoResultService;
+import lotto.service.LottoSalesService;
 
 public class Application {
     public static void main(String[] args) {
         AppConfig appConfig = new AppConfig(
-                new LottoService()
+                new LottoSalesService(),
+                new LottoResultService()
         );
 
         LottoController lottoController = appConfig.lottoController();
